payment-service/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
drf-spectacular>=0.26.0
psycopg2-binary>=2.9.9
django-cors-headers>=4.3.0
django-filter>=23.5
python-dotenv>=1.0.0
Pillow>=10.0.0
sendgrid>=6.11.0
gunicorn>=21.2.0
whitenoise>=6.6.0
dj-database-url>=2.1.0
cloudinary>=1.41.0
django-cloudinary-storage>=0.3.0


pika==1.3.2
requests==2.31.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
